[ALPS06151607] Disp: fix coverity scanned code defect
[Detail]
Fix 1 coverity scanned code defect

MTK-Commit-Id: d8a8d22623e8cf06d7e24d65eee4441c37ddfb9d

CR-Id: ALPS06151607
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6768/dispsys/ddp_disp_bdg.c
@@ -915,7 +915,8 @@ int bdg_mipi_tx_dphy_clk_setting(enum DISP_BDG_ENUM module,
 			DSI_OUTREGBIT(cmdq, struct MIPI_TX_PHY_SEL0_REG,
 					MIPI_TX_REG->MIPI_TX_PHY_SEL0,
 					MIPI_TX_PHY0_SEL,
-					pad_mapping[swap_base[MIPITX_PHY_LANE_0]]);
+					pad_mapping[(unsigned int)swap_base[
+					(unsigned int)MIPITX_PHY_LANE_0]]);
 			DSI_OUTREGBIT(cmdq, struct MIPI_TX_PHY_SEL0_REG,
 					MIPI_TX_REG->MIPI_TX_PHY_SEL0,
 					MIPI_TX_PHY1AB_SEL,
